move the logic for retrieving particular person to controllers/people.js and update files as needed
